Refactoring based on suggestions by @superseacat [#44] Moved static nested classes to upper level. Added annotations to functional interfaces. Use String.split instead of a utility class. Set Sonar to exclude tests completely (might revert this in the near future).
