+++ sonar-project.properties
@@ -5,4 +5,4 @@ sonar.analysis.mode=publish
 sonar.java.binaries=build/classes
 sonar.java.libraries=build/libs
 
-sonar.coverage.exclusions=src/test
+sonar.exclusions=src/test
