+++ src/main/java/fi/aalto/cs/intellij/notifications/Notifier.java
@@ -5,6 +5,7 @@ import com.intellij.openapi.project.Project;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+@FunctionalInterface
 public interface Notifier {
   void notify(@NotNull Notification notification, @Nullable Project project);
 }
