+++ src/test/java/fi/aalto/cs/intellij/notifications/NotificationsUtilityTest.java
@@ -1,6 +1,6 @@
 package fi.aalto.cs.intellij.notifications;
 
-import fi.aalto.cs.intellij.common.BuildInfo;
+import fi.aalto.cs.intellij.common.Version;
 import org.hamcrest.CoreMatchers;
 import org.junit.Assert;
 import org.junit.Test;
@@ -9,8 +9,8 @@ public class NotificationsUtilityTest {
 
   @Test
   public void testBetaVersionWarning() {
-    APlusNotifications.BetaVersionWarning notification =
-        new APlusNotifications.BetaVersionWarning(new BuildInfo.Version(0, 22, 315));
+    BetaVersionWarning notification =
+        new BetaVersionWarning(new Version(0, 22, 315));
 
     Assert.assertEquals("0.22.315", notification.getVersion().toString());
     Assert.assertEquals("A+", notification.getGroupId());
